dddddd

+++ canbushandling.go
@@ -115,7 +115,7 @@ func canPublish(cf CAN.CANFrame) {
 		log.Fatal(err)
 	}
 	if cf.ID > 0x7FF {
-		cf.ID &= ^(0x8000000)
+		cf.ID &= ^(uint32(0x8000000))
 	}
 
 	canSubscribe(cf.ID)
